+++ NEWS.md
@@ -6,6 +6,10 @@ title: "NEWS for the emmeans package"
   * Fix to logic error in `regrid()` (#287, revisited)
   * Fix to `nbasis` calculation in ordinal models (#387)
   * Bias-adjustment example added when we have random slopes
+  * Removed dependence on **xtable** package. The `xtable` methods are now
+    dynamically registered. This reduces the number of package dependencies
+    from 8 to 7 (as of this version).
+    
 
 ## emmeans 1.8.2
   * Extended `cld()` so it can show findings rather than non-findings,