+++ NEWS.md
@@ -2,7 +2,7 @@
 title: "NEWS for the emmeans package"
 ---
 
-## emmeans 1.8.2-080xxxx
+## emmeans 1.8.3
   * Fix to logic error in `regrid()` (#287, revisited)
   * Fix to `nbasis` calculation in ordinal models (#387)
   * Bias-adjustment example added when we have random slopes
@@ -14,7 +14,8 @@ title: "NEWS for the emmeans package"
   * Added alt text to all pictures in vignettes (#389). This makes
     the materials more accessible per guidelines from the 
     [A11Y project](https://www.a11yproject.com/).
-  * Added `"atanh"` to the options in `make.tran()`
+  * Added `"atanh"` to the options in `make.tran()` and to the
+    "named" response transformations that are auto-detected
   * `make.tran()` replaces `param` argument with `alpha` and `beta`
     (`param` is still supported for backward compatibility)
     and documentation has been revised in hopes of making everything clearer